Fix RP2 timing log build error
Cast int32_t values to int in the readIniIntRange error path to
resolve ambiguous log_raw() overload resolution.  The logging
subsystem provides overloads for int, uint8_t, uint32_t, and
uint64_t, but not int32_t, which on this platform may alias to a
distinct type causing a build failure.

+++ lib/ZuluSCSI_platform_RP2MCU/custom_timings.cpp
@@ -38,7 +38,7 @@ static bool readIniIntRange(const char *section, const char *key, int32_t curren
     if (value < min_value || value > max_value)
     {
         logmsg("Invalid value for [", section, "] ", key, " in ", CUSTOM_TIMINGS_FILE,
-            ": ", value, " (expected ", min_value, "..", max_value, ")");
+            ": ", (int)value, " (expected ", (int)min_value, "..", (int)max_value, ")");
         return false;
     }
